requirements.txt
alembic
blinker
click
Flask
Flask-JWT-Extended
Flask-Migrate
Flask-SQLAlchemy
greenlet
itsdangerous
Jinja2
Mako
MarkupSafe
PyJWT
PyMySQL
python-dotenv
SQLAlchemy
typing_extensions
Werkzeug
cryptography
